Refactored GUI.py. See notes below.
- Added snippets.py to repository for future reference.
- Removed def_create_empty_board(self). Has been moved to snippets.py.
- def get_button_from_mouse_pos(self, mouse_pos) no longer static- now iterates through self.board as opposed to board.board.
- init method contains margin size- also reorganized otherwise board properties.
